backend/requirements.txt
beautifulsoup4==4.6.0
cachetools==2.0.0
funcsigs==1.0.2
functools32==3.2.3.post2
googleapis-common-protos==1.3.5
httplib2==0.10.3
mock==2.0.0
oauth2client==4.1.2
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.5.2
PyYAML==3.12
requests==2.3
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.10.0
waitress==1.1.0
webapp2==2.5.2
WebOb==1.7.3
WebTest==2.0.27
jinja2==2.10
pyfcm==1.4.5
enum34>=1.1.6